requirements.txt
alembic==1.7.3
aniso8601==9.0.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.11
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==4.0.1
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
psycopg2-binary==2.9.3
pycparser==2.20
PyJWT==2.1.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.25
urllib3==1.26.8
Werkzeug==2.0.1